backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.1

# OCR Engine
paddlepaddle==2.6.2
paddleocr==2.7.3
numpy<2.0.0

# File Upload & Processing
python-multipart==0.0.9
opencv-python-headless==4.10.0.84

# Data Processing
numpy==1.26.4
pandas==2.2.2

# OCR Dependencies
lxml==5.3.0
openpyxl==3.1.5
Pillow==10.4.0

# Additional Utilities
psutil==6.0.0
python-dateutil==2.9.0.post0

# Development & Testing
pytest==8.3.3
httpx==0.27.2
requests==2.32.3
urllib3<2.0.0
imagekitio==3.0.0
redis==5.0.8
upstash-redis==1.5.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1

# Optional: GPU Support (if needed)
# paddlepaddle-gpu>=2.6.0
